data.js now contains domains for providers, this is used for filtering providers based on domain.
the drop down menu of filter by author now shows all authors in the
domain when the resource of the unit is selected. If not it shows the
authors of the previously selected domain. [previously it was wrong, it
showed all the authors of the activities of the resource in the course,
while the available list had different authors.]

the show my course button is disabled now, since there is no login.
Also, the function of show my course is now correct. when loading the
page for the first time, it shows the list of all courses and the button
is not selected.